Calculate social cost of carbon analytically (#54)
* 📦  Add a recent version of TidierData

To do that, I had to remove the Inequality package, because
it only allowed old versions of TidierData to be installed.
I copy-pasted the `gini` functions from Inequality into the
quantile_recycle component.

If we end up not using the Gini coefficient, we can get rid
of these functions.

* 🐥  Separate `v` from `utility` in welfare module

* 🔥  Add `co2_pulse` parameter

* 🏳️  Edit whitespace for consistency

* 📦  Add TidierFiles

* ➡️  Move a function to its module file

`country_quantile_distribution` is only called in
quantile_recycle.jl, so it makes more sense to have it in that
file, rather than in helper_functions.jl.

* 💲  Get SCC decomposition (WIP)

A first working version for a single set of parameters.

* ℹ️  Add docstrings

* 🍡  Get SCC decomposition for a vector of gammas

* Move l/10 to the function get_model_data

* ⏰  Exclude rows that are before pulse year

* Convert units of measurement for dollars and people
Convert from thousand dollars and thousand people to dollars and people

* 💥  Use damages to consumption directly

This avoids capturing abatement cost and tax redistribution
when calculating marginal damages.

* 🚧  Move development code to another file

* 🕵️  Investigate the effect of gamma (WIP)

* 🥇  Check that the average equity weight is 1

* ⚖️  Rename and rewrite a function

`country_quantile_distribution`
--> `rescale_distribution`

* 🔍  Check average marginal damages

In a given country, average marginal damages should not change
with gamma. Here, they change a little bit, which is
suspicious, but probably not enough to cause a large decrease in
SCC.

* 🧮  Check that equity weights shift as expected

Equity weights shift towards poorer deciles, as expected

* 🪲  Move ρ parameter to another function

* 🔍  Decompose results by decile

* 🔍  Compare two SCC formulas and find the problem

* ⚖️  Make SCC formula comparable across scenarios

* Use a simpler formula
* Use a reference marginal utility that doesn't
  change with gamma: the marginal utility of consumption
  evaluated at present average consumption and environment.

* 🙅  Drop rows with infinite marginal utility of E

That's four small countries with E = 0.

* 🧹  Remove debugging code

* 📊  Plot SCC decomposition

* 🧹  Remove development file

* 💬  Add a comment

Also make the code more explicit about which countries we're excluding.

---------

Co-authored-by: iroliva <iroliva@ucdavis.edu>