Make line charts of the effect of inequality (#70)
Make line charts of the effect of country-level inequality on the SCC
Facet for different values of η and θ

#### Commits
* 🔧  Add development code (work in progress)

* 🎛️  Add `additional_parameters` keyword argument

* 🐝  Remove `update_leftover_params!`

This function has unintended effects.
I think what happened is that `update_leftover_params!`
only sets values for leftover *unshared* parameters.
So a lot of shared parameters were not getting their
default values. Somehow the model was running, but it
messed up the outputs.

https://www.mimiframework.org/Mimi.jl/stable/howto/howto_9/#update_leftover_params!-1

* 🧹  Remove duplicate line and improve a comment

* 📈  Plot SCC parts side by side instead of stacked

* 📝  Update docstrings

* 💠  Get SCC decomposition for an η × θ grid

* 💠  Make facet plot for different η's and θ's

* 🛠️  Start developing temperature sensitivity check